requirements.txt
kafka-python
pyspark
faker
